# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Public Notice Portal API",
    "version": "1.0.0",
    "description": "RESTful API for managing public notices, representations, councils, and licensing applications in the UK. This API serves councils, firms, and the general public.",
    "contact": {
      "name": "API Support",
      "email": "support@publicnoticeportal.uk"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5174/api",
      "description": "Local development server"
    },
    {
      "url": "https://api.publicnoticeportal.uk",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/representations": {
      "post": {
        "summary": "Submit representation",
        "description": "Submit a public representation (objection, support, or comment) on a notice. No authentication required.",
        "tags": ["Representations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["noticeId", "submitterName", "submitterEmail", "type", "content"],
                "properties": {
                  "noticeId": { "type": "string", "format": "uuid" },
                  "submitterName": { "type": "string", "minLength": 1 },
                  "submitterEmail": { "type": "string", "format": "email" },
                  "submitterPhone": { "type": "string" },
                  "submitterAddress": { "type": "string" },
                  "type": { "type": "string", "enum": ["support", "objection", "comment"] },
                  "content": { "type": "string", "minLength": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Representation submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "representation": { "$ref": "#/components/schemas/Representation" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or deadline passed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/notices/{noticeId}/representations": {
      "get": {
        "summary": "Get notice representations",
        "description": "Retrieve all representations for a specific notice. Requires authentication.",
        "tags": ["Representations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "noticeId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "User ID to check read status",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of representations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Representation" }
                }
              }
            }
          }
        }
      }
    },
    "/representations/{representationId}/mark-read": {
      "post": {
        "summary": "Mark representation as read",
        "description": "Mark a representation as read by a user. Idempotent operation. Requires authentication.",
        "tags": ["Representations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "representationId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId"],
                "properties": {
                  "userId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "representationId": { "type": "string" },
                    "alreadyRead": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token from authentication"
      }
    },
    "schemas": {
...
      "Representation": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "notice_id": { "type": "string", "format": "uuid" },
          "submitter_name": { "type": "string" },
          "submitter_email": { "type": "string", "format": "email" },
          "submitter_phone": { "type": "string", "nullable": true },
          "submitter_address": { "type": "string", "nullable": true },
          "type": { "type": "string", "enum": ["support", "objection", "comment"] },
          "content": { "type": "string" },
          "status": { "type": "string", "enum": ["submitted", "under_review", "acknowledged", "rejected"] },
          "submitted_at": { "type": "string", "format": "date-time" },
          "internal_notes": { "type": "array", "items": { "type": "object" } },
          "is_read": { "type": "boolean", "description": "Only present when userId query param provided" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    },
...
  }
}